refactor: GroundReadPort 적용 및 변경된 메서드 명으로 수정
