+++ src/main/java/com/mobble/mobbleserver/application/club/core/service/ClubQueryService.java
@@ -3,10 +3,12 @@ package com.mobble.mobbleserver.application.club.core.service;
 import com.mobble.mobbleserver.application.club.core.port.provided.ClubQueryPort;
 import com.mobble.mobbleserver.application.club.core.port.required.ClubReadPort;
 import com.mobble.mobbleserver.application.clubMember.port.required.ClubMemberReadPort;
+import com.mobble.mobbleserver.application.ground.required.GroundReadPort;
 import com.mobble.mobbleserver.application.member.port.required.MemberReadPort;
 import com.mobble.mobbleserver.domain.club.core.Club;
 import com.mobble.mobbleserver.domain.clubMember.ClubMember;
 import com.mobble.mobbleserver.domain.clubMember.ClubMemberRole;
+import com.mobble.mobbleserver.domain.ground.Ground;
 import com.mobble.mobbleserver.domain.member.Member;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.club.ClubErrorCode;
@@ -15,14 +17,12 @@ import com.mobble.mobbleserver.infrastructure.persistence.club.core.projection.C
 import com.mobble.mobbleserver.infrastructure.web.club.core.dto.request.ClubSearchRequestDto;
 import com.mobble.mobbleserver.infrastructure.web.club.core.dto.response.ClubResponseDto;
 import com.mobble.mobbleserver.infrastructure.web.club.core.dto.response.ClubSummaryDto;
+import com.mobble.mobbleserver.infrastructure.web.ground.dto.response.GroundResponseDto;
 import com.mobble.mobbleserver.refactor.adress.entity.Address;
 import com.mobble.mobbleserver.refactor.club.ageGroup.entity.AgeGroup;
 import com.mobble.mobbleserver.refactor.club.ageGroup.entity.AgeGroupType;
 import com.mobble.mobbleserver.refactor.club.ageGroup.repository.AgeGroupRepository;
 import com.mobble.mobbleserver.refactor.club.clubGround.repository.ClubGroundRepository;
-import com.mobble.mobbleserver.refactor.ground.dto.response.GroundResponseDto;
-import com.mobble.mobbleserver.refactor.ground.entity.Ground;
-import com.mobble.mobbleserver.refactor.ground.repository.GroundRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -38,9 +38,9 @@ public class ClubQueryService implements ClubQueryPort {
     private final ClubReadPort clubReadPort;
     private final MemberReadPort memberReadPort;
     private final ClubMemberReadPort clubMemberReadPort;
+    private final GroundReadPort groundReadPort;
 
     private final AgeGroupRepository ageGroupRepository;
-    private final GroundRepository groundRepository;
     private final ClubGroundRepository clubGroundRepository;
 
     @Override
@@ -88,7 +88,7 @@ public class ClubQueryService implements ClubQueryPort {
                 .map(cg -> cg.getGround().getCode())
                 .collect(Collectors.toList());
 
-        List<GroundResponseDto> groundList = groundRepository.findAllByCodeIn(groundCodes)
+        List<GroundResponseDto> groundList = groundReadPort.findAllById(groundCodes)
                 .stream()
                 .map(GroundResponseDto::toDto)
                 .toList();
